AI_Powered_LLM_RAG_Chatbot/requirements.txt
# requirements.txt
# langchain==0.1.0
# openai==1.12.0
# google-generativeai==0.3.0
# faiss-cpu==1.7.4
# python-dotenv==1.0.0
# beautifulsoup4==4.12.2
# pypdf==3.17.1

# requirements.txt
langchain==0.1.0
transformers==4.36.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
torch==2.1.2
accelerate==0.26.1
beautifulsoup4==4.12.2
pypdf==3.17.1